requirements.txt
anyio==3.6.2
async-timeout==4.0.2
cassandra-driver==3.25.0
click==8.1.3
fastapi==0.89.1
geomet==0.2.1.post1
h11==0.14.0
idna==3.4
pydantic==1.10.4
pytz==2022.7.1
redis==4.4.2
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
